+++ src/FreeElephants/RestDaemon/HttpDriver/HttpDriverInterface.php
@@ -2,14 +2,22 @@
 
 namespace FreeElephants\RestDaemon\HttpDriver;
 
+use Aerys\Host;
 use FreeElephants\RestDaemon\Endpoint\EndpointInterface;
 use FreeElephants\RestDaemon\Middleware\Collection\EndpointMiddlewareCollectionInterface;
+use Ratchet\App;
 
 /**
  * @author samizdam <samizdam@inbox.ru>
  */
 interface HttpDriverInterface
 {
+
+    /**
+     * @internal be careful: it's strong dependency without cross-vendor adapting: pre-configured Ratchet or Aerys instance
+     * @return App|Host
+     */
+    public function getRawInstance();
 
     /**
      * @param HttpServerConfig $config